refactor: deduplicate shared types and functions between rar and sevenzip packages
Extract identical Content, NestedSource structs, error sentinels, and helper
functions (GetContentSegmentCount, GetContentSegments, ValidateSegmentIntegrity,
HasExtension) into archive/common.go. Both sub-packages now use type aliases and
delegate to the shared implementations, eliminating ~200 lines of duplication.

Also exports RAR regex patterns (PartPattern, RPattern, NumericPattern) so the
sevenzip package can import them instead of duplicating, and caches the extension
map in IsAllowedFile to avoid redundant allocations per call.
